+++ CHANGELOG.md
@@ -11,18 +11,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - DP-5286 - Add permissions to allow users with the Tester role to add and edit Regulation pages.
 - DP-3872 - Display overview field on front end of Service Detail pages.
 - DP-5154 - Add a view that allows the Content team to see a list of Legacy Redirects and related info about them. 
+- DP-5020 - Updated Event Associated pages (formerly Parent) to allow Decision, News, Advisory, Regulation and Executive Order entities.
+- DP-3631 - Added upcoming events to Locations.
 
 ### Changed
 
 - DP-4420 - Updates markers in maps on Service Pages and Organizational Landing Pages, to show linked node titles of corresponding Location pages.
 - DP-5235 - Change the Featured Topics field on Org Pages to allow adding pages of the type 'Topic Page' rather than the old, no longer used 'Topic' content type.
+- DP-5020 - Changed Parent field label to Associated pages.
+- DP-5022 - Changes upcoming events logic on advisory, decision, news, executive order and regulation.
+- DP-5020 - Hide upcoming events field from decisions, advisory and executive orders.
 
 ### Removed
 
 ## [0.31.0] - August 24, 2017
 
 ### Added
-- DP-5154 Create a view visible to content administrators with 2 displays (to show or export) redirects of content, filterable by environment, Top content, title and user org. Visible at admin/redirects-data and admin/redirects-data-export.csv.
 - DP-4590 - Add new field_video ('Video') entity reference to Video Paragraph (Inline Entity Form widget). This references a Video bundle. Add new field_video_description ('Transcript and Video Description') Text field to Video bundle. Add new, themed, media--media-video.html.twig template.
 - DP-4719 - Add short description to service page front-end.
 - DP-4584 - Adds custom metatags for parent topic on service pages, news type and date on news, advisory type on advisory pages, decision type on decision pages and date on regulation pages.
@@ -62,6 +66,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - DP-4592 - Add the more info link field to the Contact information content type, which enables authors to add an organization, location or service page.
 
 ### Removed
+- DP-5020 - Removed upcoming events field from decisions, advisory and executive orders.
 
 ## [0.30.0] - August 17, 2017